+++ libpackdude/metadata.h
@@ -15,7 +15,7 @@ typedef struct {
 	unsigned int count;
 } metadata_t;
 
-bool metadata_parse(const char *raw, const uint16_t size, metadata_t *metadata);
+bool metadata_parse(const char *raw, const size_t size, metadata_t *metadata);
 bool metadata_add(metadata_t *metadata, const char *name, const char *value);
 const char *metadata_get(const metadata_t *metadata, const char *name);
 void metadata_close(metadata_t *metadata);
